refactor: extract state Ctx methods into ctx_state.go
